FallBack RAG/requirements.txt
streamlit>=1.30.0
langchain>=0.2.6
langchain-core>=0.2.6
langchain-community>=0.2.6
langchain-openai>=0.1.8
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.12.0
pypdf>=4.0.0
serpapi>=0.1.5
requests>=2.31.0